Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
google-generativeai==0.7.2
httpx==0.27.0
chromadb==0.5.4
sentence-transformers==2.7.0